Extracted internals into Internal (StrongPath).
